Enhance gRPC streaming test to send multiple requests
